esg-sustainability-app/backend/requirements.txt
fastapi
uvicorn
pandas
langchain
langchain-aws
boto3
python-dotenv
pydantic
